requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
distlib==0.3.8
Django==4.1.7
django-otp==1.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
idna==3.4
platformdirs==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.6.0
pytz==2023.2
requests==2.28.2
sqlparse==0.4.3
twilio==7.17.0
tzdata==2023.2
urllib3==1.26.15
virtualenv==20.25.0
